Fix #5
"Don't put a space after the mention in `msg.reply`"

So, we can't really customize how Serenity's Message::reply works, so
instead we made a trait that has a method called `reply_err` which does
formatting how I want it to.
The impl for reply_err is more or less stolen from Serenity.

Also make all the error strings fit with the new reply method.
(Hopefully)