Refactor a bunch of precondition logic into `#[check]`s
Create two serenity `#[check]`s that abstract away a bunch of repeated
command precondition logic so that it is not repetitive and avoids
errors or missing checks.

Add two macros `expect_game` and `expect_game_mut` that take a `ctx.data`
and attempt to get a `Game` from it, logging an error if not possible.

Migrate as many commands as possible to these new things.

Furthermore, this fixes a number of bugs where commands didn't actually
check for specific invalid `GameState`s and reduces indentation levels
in several functions.

:^)