Rewrite the Game state machine
This is a rewrite of the old FSM to follow some of the patterns presented
in https://hoverbear.org/blog/rust-state-machine-pattern/. The types
and traits should be more documented too, though not fully.

In theory this will offer no behavioural change...
In _practice_, idk, I haven't tested this.

Also fixes a bug where you could !start an already started game.